# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Referral API",
        "description": "API for managing referrals between business units",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        },
        {
            "url": "http://mytotalhealth.com.my/referral-api",
            "description": "MyHealth Server (Staging)"
        }
    ],
    "paths": {
...
        "/api/formDetails/create": {
            "post": {
                "tags": [
                    "Form Details"
                ],
                "summary": "Create form details of a form in a business unit",
                "description": "Create details for a specific form, including fields like field name, field type, and required status.",
                "operationId": "9de1b04dcbab9cd91772760a77f8d83c",
                "requestBody": {
                    "description": "Form details to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "form_id",
                                    "form_details"
                                ],
                                "properties": {
                                    "form_id": {
                                        "description": "The ID of the form",
                                        "type": "integer"
                                    },
                                    "form_details": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "field_name",
                                                "field_type",
                                                "is_required"
                                            ],
                                            "properties": {
                                                "field_name": {
                                                    "description": "Name of the field",
                                                    "type": "string"
                                                },
                                                "field_type": {
                                                    "description": "Type of the field (e.g., text, number, etc.)",
                                                    "type": "string"
                                                },
                                                "is_required": {
                                                    "description": "Indicates if the field is required",
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Form details created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Form details created successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create form details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create form details."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message from exception"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/formDetails/{formDetail}": {
            "get": {
                "tags": [
                    "Form Details"
                ],
                "summary": "Get specific form detail",
                "operationId": "dcf18ae4b7c80c488130a9bab1d8e0c6",
                "parameters": [
                    {
                        "name": "formDetail",
                        "in": "path",
                        "description": "Form detail ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form detail retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "form_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "field_name": {
                                                    "type": "string",
                                                    "example": "patient_notes"
                                                },
                                                "field_type": {
                                                    "type": "string",
                                                    "example": "textarea"
                                                },
                                                "is_required": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "field_value": {
                                                    "type": "string",
                                                    "example": "Some default text"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve form",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to retrieve form."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Form Details"
                ],
                "summary": "Delete a specific form detail",
                "operationId": "308106e20f364073ac47e28b184ede2c",
                "parameters": [
                    {
                        "name": "form",
                        "in": "path",
                        "description": "Form detail ID to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Form deleted successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete form",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to delete form."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Exception message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/formDetails/{id}": {
            "put": {
                "tags": [
                    "Form Details"
                ],
                "summary": "Update form details",
                "description": "Updates a specific form detail entry with new values.",
                "operationId": "updateFormDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the form detail",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Form data to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "form_id",
                                    "field_name",
                                    "field_type",
                                    "is_required",
                                    "field_value"
                                ],
                                "properties": {
                                    "form_id": {
                                        "description": "The ID of the business unit",
                                        "type": "integer"
                                    },
                                    "field_name": {
                                        "description": "The label name of the form",
                                        "type": "string"
                                    },
                                    "field_type": {
                                        "description": "Whether the form is hidden or not",
                                        "type": "boolean"
                                    },
                                    "is_required": {
                                        "description": "The ID of the business unit",
                                        "type": "integer"
                                    },
                                    "field_value": {
                                        "description": "The label name of the form",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Form updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Form updated successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Update failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Update failed"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Exception message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token in the format: Bearer {token}"
            }
        }
    },
...
}